style(BaseIConContainer): rename base component 'IconContainer'
