+++ src/renderer/components/BaseIconContainer.vue
@@ -1,11 +1,11 @@
 <template>
-    <div>
-        <svg :class="hoverState">
-            <use class='default' v-bind="{'xlink:href': `#${type}-${finalState}-${finalEffect}`}"></use>
-            <use class="hover" v-bind="{'xlink:href': `#${type}-hover-${finalEffect}`}"></use>
-            <use class="active" v-bind="{'xlink:href': `#${type}-active-${finalEffect}`}"></use>
-        </svg>
-    </div>
+  <div>
+    <svg :class="hoverState">
+      <use class='default' v-bind="{'xlink:href': `#${type}-${finalState}-${finalEffect}`}"></use>
+      <use class="hover" v-bind="{'xlink:href': `#${type}-hover-${finalEffect}`}"></use>
+      <use class="active" v-bind="{'xlink:href': `#${type}-active-${finalEffect}`}"></use>
+    </svg>
+  </div>
 </template>
 
 <script>
