+++ src/renderer/components/LandingView/Playlist.vue
@@ -45,7 +45,7 @@
 
 <script>
 import path from 'path';
-import Icon from '../IconContainer';
+import Icon from '../BaseIconContainer';
 export default {
   name: 'playlist',
   components: { Icon },
