+++ src/renderer/components/PlayingView/AdvanceControl.vue
@@ -20,7 +20,7 @@
 
 <script>
 import AdvanceControlMenuItem from './AdvanceControlMenuItem.vue';
-import Icon from '../IconContainer';
+import Icon from '../BaseIconContainer';
 export default {
   components: {
     AdvanceControlMenuItem,
