+++ src/renderer/components/PlayingView/PlayButton.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import Icon from '../IconContainer';
+import Icon from '../BaseIconContainer';
 export default {
   data() {
     return {
